+++ go/api/connection_management_cluster_commands.go
@@ -21,4 +21,12 @@ type ConnectionManagementClusterCommands interface {
 	ClientId() (ClusterValue[int64], error)
 
 	ClientIdWithOptions(routeOptions options.RouteOption) (ClusterValue[int64], error)
+
+	ClientSetName(connectionName string) (ClusterValue[string], error)
+
+	ClientSetNameWithOptions(connectionName string, routeOptions options.RouteOption) (ClusterValue[string], error)
+
+	ClientGetName() (ClusterValue[string], error)
+
+	ClientGetNameWithOptions(routeOptions options.RouteOption) (ClusterValue[string], error)
 }