+++ go/api/geospacial_commands.go
@@ -28,4 +28,64 @@ type GeoSpatialCommands interface {
 	GeoDist(key string, member1 string, member2 string) (Result[float64], error)
 
 	GeoDistWithUnit(key string, member1 string, member2 string, unit options.GeoUnit) (Result[float64], error)
+
+	GeoSearch(
+		key string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+	) ([]string, error)
+
+	GeoSearchWithInfoOptions(
+		key string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		infoOptions options.GeoSearchInfoOptions,
+	) ([]options.Location, error)
+
+	GeoSearchWithResultOptions(
+		key string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+	) ([]string, error)
+
+	GeoSearchWithFullOptions(
+		key string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+		infoOptions options.GeoSearchInfoOptions,
+	) ([]options.Location, error)
+
+	GeoSearchStore(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+	) (int64, error)
+
+	GeoSearchStoreWithInfoOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		storeInfoOptions options.GeoSearchStoreInfoOptions,
+	) (int64, error)
+
+	GeoSearchStoreWithResultOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+	) (int64, error)
+
+	GeoSearchStoreWithFullOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+		storeInfoOptions options.GeoSearchStoreInfoOptions,
+	) (int64, error)
 }