+++ go/api/options/constants.go
@@ -198,6 +198,7 @@ const (
 	ObjectTypeStream ObjectType = "stream"
 )
 
+// The unit of measurement for the geospatial data
 type GeoUnit string
 
 const (
@@ -209,4 +210,25 @@ const (
 	GeoUnitMiles GeoUnit = "mi"
 	// Represents distance in feet
 	GeoUnitFeet GeoUnit = "ft"
+)
+
+// Valkey API keywords for the `GeoSearch` command
+const (
+	WithCoordValkeyApi = "WITHCOORD"
+	WithDistValkeyApi  = "WITHDIST"
+	WithHashValkeyApi  = "WITHHASH"
+)
+
+// The search origin API keyword for the `GeoCoordOrigin`
+const (
+	GeoCoordOriginAPIKeyword  = "FROMLONLAT"
+	GeoMemberOriginAPIKeyword = "FROMMEMBER"
+)
+
+// The shape of the search area for the `GeoSearch` command
+type SearchShape string
+
+const (
+	BYRADIUS SearchShape = "BYRADIUS"
+	BYBOX    SearchShape = "BYBOX"
 )