+++ go/api/server_management_commands.go
@@ -37,4 +37,8 @@ type ServerManagementCommands interface {
 	Lolwut() (string, error)
 
 	LolwutWithOptions(opts options.LolwutOptions) (string, error)
+
+	LastSave() (int64, error)
+
+	ConfigResetStat() (string, error)
 }